Changes : - Backend Inventory CRUD (belum upload foto, search, validasi user, loading) - Ganti semua tipe data bilangan jadi Long.

+++ app/src/main/java/id/fishku/fishkuseller/login/LoginViewModel.kt
@@ -50,11 +50,11 @@ class LoginViewModel(private val pref: LoginPref): ViewModel() {
 
     }
 
-    fun getSellerId(): LiveData<Int>{
+    fun getSellerId(): LiveData<Long> {
         return pref.getSellerId().asLiveData()
     }
 
-    fun saveSellerId(sellerId: Int){
+    fun saveSellerId(sellerId: Long){
         viewModelScope.launch {
             pref.saveSellerId(sellerId)
         }
